requirements.txt
# Core dependencies for task scheduler
schedule==1.2.0
loguru==0.7.2
PyYAML==6.0.1
psutil==5.9.6
setproctitle==1.3.3

# Development dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
